+++ library/src/schemas/array/types.ts
@@ -1,4 +1,4 @@
-import type { BaseIssue, MaybeReadonly } from '../../types/index.ts';
+import type { BaseIssue } from '../../types/index.ts';
 
 /**
  * Array issue type.
@@ -16,30 +16,4 @@ export interface ArrayIssue extends BaseIssue<unknown> {
    * The expected property.
    */
   readonly expected: 'Array';
-}
-
-/**
- * Array path item type.
- */
-export interface ArrayPathItem {
-  /**
-   * The path item type.
-   */
-  readonly type: 'array';
-  /**
-   * The path item origin.
-   */
-  readonly origin: 'value';
-  /**
-   * The path item input.
-   */
-  readonly input: MaybeReadonly<unknown[]>;
-  /**
-   * The path item key.
-   */
-  readonly key: number;
-  /**
-   * The path item value.
-   */
-  readonly value: unknown;
 }