+++ library/src/schemas/map/types.ts
@@ -25,32 +25,6 @@ export interface MapIssue extends BaseIssue<unknown> {
 }
 
 /**
- * Map path item type.
- */
-export interface MapPathItem {
-  /**
-   * The path item type.
-   */
-  readonly type: 'map';
-  /**
-   * The path item origin.
-   */
-  readonly origin: 'key' | 'value';
-  /**
-   * The path item input.
-   */
-  readonly input: Map<unknown, unknown>;
-  /**
-   * The path item key.
-   */
-  readonly key: unknown;
-  /**
-   * The path item value.
-   */
-  readonly value: unknown;
-}
-
-/**
  * Infer map input type.
  */
 export type InferMapInput<