+++ library/src/schemas/record/record.ts
@@ -4,13 +4,13 @@ import type {
   Dataset,
   ErrorMessage,
   InferIssue,
+  ObjectPathItem,
 } from '../../types/index.ts';
 import { _addIssue, _isValidObjectKey } from '../../utils/index.ts';
 import type {
   InferRecordInput,
   InferRecordOutput,
   RecordIssue,
-  RecordPathItem,
 } from './types.ts';
 
 /**
@@ -138,8 +138,8 @@ export function record(
             // If there are issues, capture them
             if (keyDataset.issues) {
               // Create record path item
-              const pathItem: RecordPathItem = {
-                type: 'record',
+              const pathItem: ObjectPathItem = {
+                type: 'object',
                 origin: 'key',
                 input: input as Record<string, unknown>,
                 key: entryKey,
@@ -174,8 +174,8 @@ export function record(
             // If there are issues, capture them
             if (valueDataset.issues) {
               // Create record path item
-              const pathItem: RecordPathItem = {
-                type: 'record',
+              const pathItem: ObjectPathItem = {
+                type: 'object',
                 origin: 'value',
                 input: input as Record<string, unknown>,
                 key: entryKey,