+++ library/src/schemas/set/set.ts
@@ -4,14 +4,10 @@ import type {
   Dataset,
   ErrorMessage,
   InferIssue,
+  SetPathItem,
 } from '../../types/index.ts';
 import { _addIssue } from '../../utils/index.ts';
-import type {
-  InferSetInput,
-  InferSetOutput,
-  SetIssue,
-  SetPathItem,
-} from './types.ts';
+import type { InferSetInput, InferSetOutput, SetIssue } from './types.ts';
 
 /**
  * Set schema type.