+++ library/src/schemas/strictTuple/strictTuple.ts
@@ -1,4 +1,5 @@
 import type {
+  ArrayPathItem,
   BaseIssue,
   BaseSchema,
   Dataset,
@@ -7,7 +8,6 @@ import type {
   InferTupleIssue,
   InferTupleOutput,
   TupleItems,
-  TuplePathItem,
 } from '../../types/index.ts';
 import { _addIssue } from '../../utils/index.ts';
 import type { StrictTupleIssue } from './types.ts';
@@ -102,8 +102,8 @@ export function strictTuple(
           // If there are issues, capture them
           if (itemDataset.issues) {
             // Create tuple path item
-            const pathItem: TuplePathItem = {
-              type: 'tuple',
+            const pathItem: ArrayPathItem = {
+              type: 'array',
               origin: 'value',
               input,
               key,
@@ -154,7 +154,7 @@ export function strictTuple(
             expected: 'never',
             path: [
               {
-                type: 'tuple',
+                type: 'array',
                 origin: 'value',
                 input,
                 key: this.items.length,