+++ library/src/schemas/tuple/tupleAsync.test.ts
@@ -186,7 +186,7 @@ describe('tupleAsync', () => {
       received: '123',
       path: [
         {
-          type: 'tuple',
+          type: 'array',
           origin: 'value',
           input: [123, 456, 'true'],
           key: 0,
@@ -213,7 +213,7 @@ describe('tupleAsync', () => {
             received: '"true"',
             path: [
               {
-                type: 'tuple',
+                type: 'array',
                 origin: 'value',
                 input: input,
                 key: 2,
@@ -259,14 +259,14 @@ describe('tupleAsync', () => {
             received: '"123"',
             path: [
               {
-                type: 'tuple',
+                type: 'array',
                 origin: 'value',
                 input: input,
                 key: 0,
                 value: input[0],
               },
               {
-                type: 'tuple',
+                type: 'array',
                 origin: 'value',
                 input: input[0],
                 key: 1,
@@ -283,7 +283,7 @@ describe('tupleAsync', () => {
             received: 'null',
             path: [
               {
-                type: 'tuple',
+                type: 'array',
                 origin: 'value',
                 input: input,
                 key: 1,