+++ library/src/schemas/tupleWithRest/tupleWithRest.ts
@@ -1,4 +1,5 @@
 import type {
+  ArrayPathItem,
   BaseIssue,
   BaseSchema,
   Dataset,
@@ -10,7 +11,6 @@ import type {
   InferTupleIssue,
   InferTupleOutput,
   TupleItems,
-  TuplePathItem,
 } from '../../types/index.ts';
 import { _addIssue } from '../../utils/index.ts';
 import type { TupleWithRestIssue } from './types.ts';
@@ -124,8 +124,8 @@ export function tupleWithRest(
           // If there are issues, capture them
           if (itemDataset.issues) {
             // Create tuple path item
-            const pathItem: TuplePathItem = {
-              type: 'tuple',
+            const pathItem: ArrayPathItem = {
+              type: 'array',
               origin: 'value',
               input,
               key,
@@ -174,8 +174,8 @@ export function tupleWithRest(
             // If there are issues, capture them
             if (itemDataset.issues) {
               // Create tuple path item
-              const pathItem: TuplePathItem = {
-                type: 'tuple',
+              const pathItem: ArrayPathItem = {
+                type: 'array',
                 origin: 'value',
                 input,
                 key,