+++ library/src/types/object.ts
@@ -25,32 +25,6 @@ import type { BaseSchema, BaseSchemaAsync } from './schema.ts';
 import type { MarkOptional, MaybeReadonly, Prettify } from './utils.ts';
 
 /**
- * Object path item type.
- */
-export interface ObjectPathItem {
-  /**
-   * The path item type.
-   */
-  readonly type: 'object';
-  /**
-   * The path item origin.
-   */
-  readonly origin: 'value';
-  /**
-   * The path item input.
-   */
-  readonly input: Record<string, unknown>;
-  /**
-   * The path item key.
-   */
-  readonly key: string;
-  /**
-   * The path item value.
-   */
-  readonly value: unknown;
-}
-
-/**
  * Object entries type.
  */
 export interface ObjectEntries {