+++ library/CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to the library will be documented in this file.
 - Add `partialCheck` and `partialCheckAsync` action (issue #145, #260)
 - Rename `every` and `some` action to `everyItem` and `someItem`
 - Rename `_isAllowedObjectKey` to `_isValidObjectKey` and add check for inherited properties
+- Remove `RecordPathItem` and `TuplePathItem` type and refactor code
 - Fix `received` property of issue in `date` schema for invalid dates (issue #654)
 
 ## v0.32.0 (June 14, 2024)